src/requirements.txt
# Core Dependencies
numba>=0.56.0
numpy>=1.21.0
codecarbon>=2.1.0
virtualenv>=20.0.0
pandas>=1.3.0

# Machine Learning & Vision
torch>=2.0.0
torchvision>=0.15.0
tokenizers>=0.13.0
Pillow>=9.0.0
tqdm>=4.65.0

# Testing & Development
pytest>=7.0.0
sniffio>=1.2.0
exceptiongroup>=1.1.0

# Environment Variables
python-dotenv>=0.21.0